rest of the features + more coherent management of reinvite across the app
