+++ app/javascript/controllers/tooltip_controller.js
@@ -20,24 +20,21 @@ export default class extends Controller {
   csvExportUsers() {
     tippy(this.element, {
       content:
-        "Les usagers seront exportés en fonction du contexte et des éventuels filtres sélectionnés",
-      placement: "bottom",
+        "Les usagers correspondant aux filtres actuels seront exportés",
     });
   }
 
   csvExportParticipations() {
     tippy(this.element, {
       content:
-        "Les RDVs seront exportés en fonction du contexte et des éventuels filtres sélectionnés",
-      placement: "bottom",
+        "Les RDVs des usagers correspondant aux filtres actuels seront exportés",
     });
   }
 
   batchActions() {
     tippy(this.element, {
       content:
-        "Les usagers présentés correspondront aux filtres choisis sur cet écran",
-      placement: "bottom",
+        "Les usagers présentés correspondront aux filtres actuels",
     });
   }
