+++ app/javascript/stylesheets/components/_form.scss
@@ -87,3 +87,11 @@
   width: 40px;
   cursor: pointer;
 }
+
+.checkbox-margin-left {
+  margin-left: 10px;
+}
+
+.fa-redo-alt {
+  font-size: 0.9em;
+}
