+++ app/views/users/_users_list_header.html.erb
@@ -144,7 +144,7 @@
     </div>
     <% if @current_configuration.present? %>
       <div class="d-flex flex-column align-items-end dropdown">
-        <%= link_to(new_structure_batch_action_path(**params.to_unsafe_h), class: "btn btn-blue") do %>
+        <%= link_to(new_structure_batch_action_path(**params.to_unsafe_h), class: "btn btn-blue", data: { controller: "tooltip", action: "mouseover->tooltip#batchActions"}) do %>
           Envoyer des invitations en masse
         <% end %>
       </div>
