+++ app/views/users/_users_table_for_motif_category.html.erb
@@ -26,7 +26,7 @@
             <% if rdv_context.rdv_pending? || rdv_context.closed? %>
               <!-- show nothing -->
             <% elsif user.can_be_invited_through?(invitation_format) %>
-              <%= render "invitations/checkbox_form", user: user, invitation_format: invitation_format, motif_category: @current_motif_category, rdv_context: rdv_context, checked: false, disabled: user.deleted? %>
+              <%= render "invitations/checkbox_form", user: user, invitation_format: invitation_format, motif_category: @current_motif_category, rdv_context: rdv_context, checked: checked = rdv_context.first_invitation_relative_to_last_participation_by(invitation_format).present?, disabled: checked || user.deleted? %>
             <% else %>
               -
             <% end %>
