+++ app/views/users/batch_actions/new.html.erb
@@ -1,5 +1,11 @@
 <%= react_component(
-  "pages/UsersBatchActions", {
-    usersFromApp: @users, organisation: @organisation, configuration: @current_configuration, department: @department, motifCategoryName: @motif_category_name, currentAgent: current_agent
+  "pages/users/BatchActions", {
+    usersFromApp: @users,
+    organisation: @organisation,
+    configuration: @current_configuration,
+    department: @department,
+    motifCategoryName: @motif_category_name,
+    currentAgent: current_agent,
+    backToUsersListUrl: @back_to_users_list_url
   }
 ) %>
