+++ app/views/users/uploads/new.html.erb
@@ -1,5 +1,12 @@
 <%= react_component(
-  "pages/UsersUpload", {
-    organisation: @organisation, configuration: @current_configuration, columnNames: @file_configuration.column_attributes, sheetName: @file_configuration.sheet_name, department: @department, motifCategoryName: @motif_category_name, currentAgent: current_agent, tags: (@organisation || @department).tags
+  "pages/users/Uploads", {
+    organisation: @organisation,
+    configuration: @current_configuration,
+    columnNames: @file_configuration.column_attributes,
+    sheetName: @file_configuration.sheet_name,
+    department: @department,
+    motifCategoryName: @motif_category_name,
+    currentAgent: current_agent,
+    tags: (@organisation || @department).tags
   }
 ) %>
